fix: upgrade to mui version that supports rsc
