+++ samples/msal-node-samples/nextjs-rsc-ts/README.md
@@ -133,7 +133,7 @@ The sample uses `createCookieSessionStorage` from `@remix-run/node`, which might
 
 Session lifetime has not been handled at all, and the RedisCachePlugin does not evict anything, this should be reviewed when implementing in your own applications.
 
-The `./src/components/mui.tsx` re-exports some components from Material UI, and marks them as Client Components. This allows them to be used from Server Components.
+Material UI was integrated by following the [App Router guide](https://mui.com/material-ui/guides/next-js-app-router/).
 
 The `middleware.ts` currently exposes the currently accessed url so that it can be retrieved in server components. This allows you to send the user back to where they originally were dynamically. If your application is hosted behind a proxy, you'd have to compute the url based on [forwarded headers](https://developer.mozilla.org/en-US/docs/Web/HTTP/Headers/Forwarded), depending on what your hosting platform exposes.
