+++ samples/msal-node-samples/nextjs-rsc-ts/package.json
@@ -10,17 +10,17 @@
   },
   "dependencies": {
     "@azure/msal-node": "1.16",
-    "@emotion/react": "^11.11.0",
+    "@emotion/react": "^11.11.1",
     "@emotion/styled": "^11.11.0",
-    "@mui/icons-material": "^5.11.16",
-    "@mui/material": "^5.13.3",
+    "@mui/icons-material": "^5.14.0",
+    "@mui/material": "^5.14.0",
     "@remix-run/node": "^1.16.1",
     "@types/node": "20.1.1",
     "@types/react": "^18.2.8",
     "@types/react-dom": "18.2.4",
     "eslint": "8.40.0",
     "eslint-config-next": "13.4.1",
-    "next": "^13.4.4",
+    "next": "^13.4.9",
     "react": "18.2.0",
     "react-dom": "18.2.0",
     "redis": "^4.6.6",
