+++ samples/msal-node-samples/nextjs-rsc-ts/src/components/ConsentRouteHandlerButton.tsx
@@ -1,4 +1,4 @@
-import { Button } from "./mui";
+import Button from "@mui/material/Button";
 
 // calls src/app/auth/eventConsent/route.ts
 export default function ConsentRouteHandlerButton() {
