+++ samples/msal-node-samples/nextjs-rsc-ts/src/components/ConsentServerActionButton.tsx
@@ -1,5 +1,5 @@
 import { acquireCalendarTokenInteractive } from "~/actions/auth";
-import { Button } from "./mui";
+import Button from "@mui/material/Button";
 
 export default function ConsentServerActionButton() {
   return (
