+++ samples/msal-node-samples/nextjs-rsc-ts/src/components/LoginServerActionButton.tsx
@@ -1,5 +1,5 @@
 import { login } from "~/actions/auth";
-import { Button } from "./mui";
+import Button from "@mui/material/Button";
 
 export default function LoginServerActionButton() {
   return (
