+++ samples/msal-node-samples/nextjs-rsc-ts/src/components/LogoutRouteHandlerButton.tsx
@@ -1,4 +1,4 @@
-import { Button } from "./mui";
+import Button from "@mui/material/Button";
 
 // calls src/app/auth/logout/route.ts
 export default function LogoutRouteHandlerButton() {
