+++ samples/msal-node-samples/nextjs-rsc-ts/src/components/ProfilePicture.tsx
@@ -1,5 +1,5 @@
 import { authProvider } from "~/services/auth";
-import { Avatar } from "./mui";
+import Avatar from "@mui/material/Avatar";
 
 export default async function ProfilePicture() {
   if (!(await authProvider.getAccount())) {
