+++ samples/msal-node-samples/nextjs-rsc-ts/src/components/theme/ThemeRegistry.tsx
@@ -1,25 +1,70 @@
+// this file is needed to make material and emotion work with nextjs app router
+// https://mui.com/material-ui/guides/next-js-app-router/#using-material-ui-with-a-custom-theme
+
 "use client";
 
-// this file is needed to make material and emotion work with nextjs app router
-// https://github.com/mui/material-ui/pull/37315
-
-import * as React from "react";
+import React, { PropsWithChildren } from "react";
+import createCache, { Options } from "@emotion/cache";
+import { useServerInsertedHTML } from "next/navigation";
+import { CacheProvider } from "@emotion/react";
 import { ThemeProvider } from "@mui/material/styles";
 import CssBaseline from "@mui/material/CssBaseline";
-import { NextAppDirEmotionCacheProvider } from "./EmotionCache";
 import { theme } from "~/styles/theme";
 
+type ThemeRegistryProps = PropsWithChildren<{ options: Options }>;
+
+// This implementation is from emotion-js
+// https://github.com/emotion-js/emotion/issues/2928#issuecomment-1319747902
 export default function ThemeRegistry({
+  options,
   children,
-}: {
-  children: React.ReactNode;
-}) {
+}: ThemeRegistryProps) {
+  const [{ cache, flush }] = React.useState(() => {
+    const cache = createCache(options);
+    cache.compat = true;
+    const prevInsert = cache.insert;
+    let inserted: string[] = [];
+    cache.insert = (...args) => {
+      const serialized = args[1];
+      if (cache.inserted[serialized.name] === undefined) {
+        inserted.push(serialized.name);
+      }
+      return prevInsert(...args);
+    };
+    const flush = () => {
+      const prevInserted = inserted;
+      inserted = [];
+      return prevInserted;
+    };
+    return { cache, flush };
+  });
+
+  useServerInsertedHTML(() => {
+    const names = flush();
+    if (names.length === 0) {
+      return null;
+    }
+    let styles = "";
+    for (const name of names) {
+      styles += cache.inserted[name];
+    }
+    return (
+      <style
+        key={cache.key}
+        data-emotion={`${cache.key} ${names.join(" ")}`}
+        dangerouslySetInnerHTML={{
+          __html: styles,
+        }}
+      />
+    );
+  });
+
   return (
-    <>
-      <CssBaseline />
-      <NextAppDirEmotionCacheProvider options={{ key: "mui" }}>
-        <ThemeProvider theme={theme}>{children}</ThemeProvider>
-      </NextAppDirEmotionCacheProvider>
-    </>
+    <CacheProvider value={cache}>
+      <ThemeProvider theme={theme}>
+        <CssBaseline />
+        {children}
+      </ThemeProvider>
+    </CacheProvider>
   );
 }
